requirements.txt
tensorflow
streamlit
opencv-python
pillow
numpy
pandas
matplotlib
scikit-learn
kagglehub
